DIS-276 prevent user logout on network error
If Discovery is unreachable or returns a non-2XX status when the background
refresh process executes (every 15 min in 25.02, every 5 in 25.03), the user
session would be invalidated and the user would have to reauthenticate.

Now, an Error is thrown, which prevents calls to reloadProfile from
executing onSuccess or .then() steps, which include the session invalidation
call.  The user profile is not updated, but the authentication session persists